+++ autoware_utils/package.xml
@@ -19,8 +19,8 @@
   <depend>autoware_internal_planning_msgs</depend>
   <depend>autoware_perception_msgs</depend>
   <depend>autoware_planning_msgs</depend>
+  <depend>autoware_utils_geometry</depend>
   <depend>autoware_utils_pcl</depend>
-  <depend>autoware_utils_tf2</depend>
   <depend>autoware_vehicle_msgs</depend>
   <depend>builtin_interfaces</depend>
   <depend>diagnostic_msgs</depend>
