+++ autoware_utils_geometry/package.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>autoware_utils_tf2</name>
+  <name>autoware_utils_geometry</name>
   <version>1.1.0</version>
-  <description>The autoware_utils_tf2 package</description>
+  <description>The autoware_utils_geometry package</description>
   <maintainer email="egon.kang@autocore.ai">Jian Kang</maintainer>
   <maintainer email="ryohsuke.mitsudome@tier4.jp">Ryohsuke Mitsudome</maintainer>
   <maintainer email="esteve.fernandez@tier4.jp">Esteve Fernandez</maintainer>
@@ -18,7 +18,6 @@
   <depend>rclcpp</depend>
   <depend>tf2_ros</depend>
 
-  <test_depend>ament_cmake_ros</test_depend>
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>autoware_lint_common</test_depend>
