+++ autoware_utils_pcl/package.xml
@@ -14,11 +14,10 @@
   <buildtool_depend>ament_cmake_auto</buildtool_depend>
   <buildtool_depend>autoware_cmake</buildtool_depend>
 
-  <depend>autoware_utils_tf2</depend>
+  <depend>autoware_utils_geometry</depend>
   <depend>pcl_conversions</depend>
   <depend>pcl_ros</depend>
 
-  <test_depend>ament_cmake_ros</test_depend>
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>autoware_lint_common</test_depend>
