--- autoware_utils_pcl/test/src/build.cpp
@@ -1,23 +1,0 @@
-// Copyright 2025 The Autoware Contributors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#include <autoware_utils_pcl/managed_transform_buffer.hpp>
-#include <autoware_utils_pcl/pcl_conversion.hpp>
-#include <autoware_utils_pcl/transforms.hpp>
-
-#include <gtest/gtest.h>
-
-TEST(AutowareUtilsPcl, Build)
-{
-}
