--- autoware_utils_tf2/CMakeLists.txt
@@ -1,14 +1,0 @@
-cmake_minimum_required(VERSION 3.14)
-project(autoware_utils_tf2)
-
-find_package(autoware_cmake REQUIRED)
-autoware_package()
-
-if(BUILD_TESTING)
-  file(GLOB_RECURSE test_files test/**/*.cpp)
-  ament_add_ros_isolated_gtest(test_${PROJECT_NAME} ${test_files})
-  ament_target_dependencies(test_${PROJECT_NAME} geometry_msgs rclcpp tf2_ros)
-  target_include_directories(test_${PROJECT_NAME} PRIVATE include)
-endif()
-
-ament_auto_package()
