--- autoware_utils_tf2/README.md
@@ -1,1 +1,0 @@
-# autoware_utils_tf2
